pyproject.toml
[tool.poetry]
name = "pseudohealthy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "1.22.0"
torch = {url = "https://download.pytorch.org/whl/cu116/torch-1.12.1%2Bcu116-cp38-cp38-linux_x86_64.whl"}
#torch = "^2.0.1"
torchvision = {url = "https://download.pytorch.org/whl/cu116/torchvision-0.13.1%2Bcu116-cp38-cp38-linux_x86_64.whl"}
#torchvision = "^0.15.1"
pydicom = "^2.3.0"
wandb = "^0.13.4"
matplotlib = "^3.6.1"
tqdm = "^4.64.1"
opencv-python = "^4.6.0.66"
pandas = "^1.5.1"
albumentations = "^1.3.0"
pynrrd = "^1.0.0"
gdcm = "^1.1"
pylibjpeg = "^1.4.0"
pylibjpeg-libjpeg = "^1.3.2"
jupyter = "^1.0.0"
notebook = "^6.5.2"
jupyterlab = "^3.5.0"
pytorch-lightning = "^1.8.0.post1"
omegaconf = "^2.2.3"
einops = "^0.6.0"
nibabel = "^4.0.2"
timm = "^0.6.12"
Cython = "^0.29.33"
torchmetrics = "^0.11.4"
ml-collections = "^0.1.1"
transformers = "^4.28.1"
map-boxes = "^1.0.5"
yacs = "^0.1.8"
MedPy = "^0.4.0"
tensorboardX = "^2.6.1"
pytorch-fid = "^0.3.0"
numba = "^0.57.1"
accelerate = "^0.21.0"
saliency = "^0.2.0"
captum = "^0.6.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
